refactor: Message => Constant로 변경 후 상수화 적용
