+++ src/BridgeGame.js
@@ -47,7 +47,6 @@ class BridgeGame {
    * 재시작을 위해 필요한 메서드의 반환 값(return value), 인자(parameter)는 자유롭게 추가하거나 변경할 수 있다.
    */
   retry(app) {
-    console.log(this.#tryNum);
     this.#moveHistory = [];
     this.#tryNum += 1;
     app.startGame();
