+++ src/BridgeMaker.js
@@ -1,3 +1,5 @@
+const { KEY } = require('./Constant');
+
 /**
  * 다리의 길이를 입력 받아서 다리를 생성해주는 역할을 한다.
  */
@@ -11,7 +13,7 @@ const BridgeMaker = {
     const bridge = [];
     for (let i = 0; i < size; i += 1) {
       const bridgeShape = generateRandomNumber();
-      bridge.push(bridgeShape === 0 ? 'D' : 'U');
+      bridge.push(bridgeShape === 0 ? KEY.DOWN : KEY.UP);
     }
     return bridge;
   },
