+++ src/BridgeRandomNumberGenerator.js
@@ -1,13 +1,10 @@
-const MissionUtils = require("@woowacourse/mission-utils");
+const MissionUtils = require('@woowacourse/mission-utils');
 
 const BridgeRandomNumberGenerator = {
   RANDOM_LOWER_INCLUSIVE: 0,
   RANDOM_UPPER_INCLUSIVE: 1,
   generate() {
-    return MissionUtils.Random.pickNumberInRange(
-      BridgeRandomNumberGenerator.RANDOM_LOWER_INCLUSIVE,
-      BridgeRandomNumberGenerator.RANDOM_UPPER_INCLUSIVE
-    );
+    return MissionUtils.Random.pickNumberInRange(BridgeRandomNumberGenerator.RANDOM_LOWER_INCLUSIVE, BridgeRandomNumberGenerator.RANDOM_UPPER_INCLUSIVE);
   },
 };
